The 10.9 Catalog is not Copied for the Root Branch When Applying Changes
When changes are made to enable/disable updates with the Root Branch
for 10.9, the changes are not copied to the correct location when
applying those updates. Modified the code to loop through all catalogs
for that branch which should work for all OS versions going forward
without changing the code.