Changing the DocumentRoot is Incorrect
Changing the DocumentRoot is not the right way to handle having the php
files in a different location. Move the php files to the correct
location instead. Without this change, the NetSUS and JDS cannot
coexist on a CentOS/RedHat server and the timezone setting doesn’t get
pulled correctly since the sed string below this change expects the
files to be in the correct location.